# Ast.List<T> performance tweaks