--- a/src/Esprima/JavascriptParser.cs
+++ b/src/Esprima/JavascriptParser.cs
@@ -1039,10 +1039,11 @@ namespace Esprima
 
                     break;
                 case Nodes.ArrayExpression:
-                    var elements = new Ast.List<ArrayPatternElement>();
-
-                    foreach (var element in expr.As<ArrayExpression>().Elements)
+                    var list = expr.As<ArrayExpression>().Elements;
+                    var elements = new Ast.List<ArrayPatternElement>(list.Count);
+                    for (var i = 0; i < list.Count; i++)
                     {
+                        var element = list[i];
                         if (element != null)
                         {
                             elements.Add(ReinterpretExpressionAsPattern(element).As<ArrayPatternElement>());
@@ -1060,13 +1061,15 @@ namespace Esprima
 
                     break;
                 case Nodes.ObjectExpression:
-                    var properties = new Ast.List<Property>();
-
-                    foreach (var property in (expr.As<ObjectExpression>()).Properties)
+                    var props = expr.As<ObjectExpression>().Properties;
+                    var properties = new Ast.List<Property>(props.Count);
+                    for (var i = 0; i < props.Count; i++)
                     {
+                        var property = props[i];
                         property.Value = ReinterpretExpressionAsPattern(property.Value).As<PropertyValue>();
                         properties.Add(property);
                     }
+
                     node = new ObjectPattern(properties);
                     node.Range = expr.Range;
                     node.Location = _config.Loc ? expr.Location : null;
@@ -1148,11 +1151,13 @@ namespace Esprima
                                     Expect("=>");
                                 }
                                 _context.IsBindingElement = false;
-                                var reinterpretedExpressions = new Ast.List<Expression>();
-                                foreach (var expression in expressions)
+                                var reinterpretedExpressions = new Ast.List<Expression>(expressions.Count);
+                                for (var index = 0; index < expressions.Count; index++)
                                 {
+                                    var expression = expressions[index];
                                     reinterpretedExpressions.Add(ReinterpretExpressionAsPattern(expression).As<Expression>());
                                 }
+
                                 expressions = reinterpretedExpressions;
                                 arrow = true;
                                 expr = new ArrowParameterPlaceHolder(expressions.Select(e => (INode) e));
@@ -1192,11 +1197,13 @@ namespace Esprima
                                 if (expr.Type == Nodes.SequenceExpression)
                                 {
                                     var sequenceExpression = expr.As<SequenceExpression>();
-                                    var reinterpretedExpressions = new Ast.List<Expression>();
-                                    foreach (var expression in sequenceExpression.Expressions)
+                                    var reinterpretedExpressions = new Ast.List<Expression>(sequenceExpression.Expressions.Count);
+                                    for (var i = 0; i < sequenceExpression.Expressions.Count; i++)
                                     {
+                                        var expression = sequenceExpression.Expressions[i];
                                         reinterpretedExpressions.Add(ReinterpretExpressionAsPattern(expression).As<Expression>());
                                     }
+
                                     sequenceExpression.Expressions = reinterpretedExpressions;
                                 }
                                 else
@@ -1737,22 +1744,28 @@ namespace Esprima
                     CheckPatternParam(options, param.As<AssignmentPattern>().Left);
                     break;
                 case Nodes.ArrayPattern:
-                    foreach (var element in param.As<ArrayPattern>().Elements)
+                    var list = param.As<ArrayPattern>().Elements;
+                    for (var i = 0; i < list.Count; i++)
                     {
+                        var element = list[i];
                         if (element != null)
                         {
                             CheckPatternParam(options, element);
                         }
                     }
+
                     break;
                 case Nodes.YieldExpression:
                     break;
                 default:
                     //assert(param.type == Nodes.ObjectPattern, 'Invalid type');
-                    foreach (var property in param.As<ObjectPattern>().Properties)
+                    var properties = param.As<ObjectPattern>().Properties;
+                    for (var i = 0; i < properties.Count; i++)
                     {
+                        var property = properties[i];
                         CheckPatternParam(options, property.Value);
                     }
+
                     break;
             }
         }
@@ -2973,15 +2986,16 @@ namespace Esprima
 
                 case TokenType.Punctuator:
                     var value = _lookahead.Value;
-                    if ((string)value == "{")
+                    var s = (string) value;
+                    if (s == "{")
                     {
                         statement = ParseBlock();
                     }
-                    else if ((string)value == "(")
+                    else if (s == "(")
                     {
                         statement = ParseExpressionStatement();
                     }
-                    else if ((string)value == ";")
+                    else if (s == ";")
                     {
                         statement = ParseEmptyStatement();
                     }