--- a/src/Esprima/JitHelper.cs
+++ b/src/Esprima/JitHelper.cs
@@ -11,5 +11,14 @@ namespace Esprima
             return default;
 #pragma warning restore 162
         }
+        
+        public static T Throw<T>() where T : Exception, new()
+        {
+            throw new T();
+#pragma warning disable 162 // unreachable code
+            return default;
+#pragma warning restore 162
+        }
+
     }
 }