I went checked what could be done to further improve the new List performance and here are some tweaks. Feel free to reject, cherry-pick or use however you want. You can also double-check the numbers whether they make sense. There's not much improvement on file parsing though, underscore remains a mystery...

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17763
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.101
  [Host]     : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT


```
|               Method |      Mean |     Error |    StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------------- |----------:|----------:|----------:|--------:|-------:|----------:|
|           AstListAdd | 23.820 us | 0.1754 us | 0.1641 us | 31.2195 |      - |  131368 B |
|           BclListAdd | 26.968 us | 0.1820 us | 0.1520 us | 31.2195 | 7.7820 |  131408 B |
|           AstListFor |  9.647 us | 0.0177 us | 0.0147 us |       - |      - |       0 B |
|           BclListFor | 11.790 us | 0.0060 us | 0.0047 us |       - |      - |       0 B |
|     AstListAddAndFor | 32.123 us | 0.1109 us | 0.0983 us | 31.1890 |      - |  131368 B |
|     BclListAddAndFor | 32.900 us | 0.1351 us | 0.1264 us | 31.1890 |      - |  131408 B |
| AstListAddAndForEach | 49.238 us | 0.1208 us | 0.1130 us | 31.1890 |      - |  131368 B |
| BclListAddAndForEach | 47.575 us | 0.2954 us | 0.2764 us | 31.1890 |      - |  131408 B |
|       AstListForEach | 25.194 us | 0.0198 us | 0.0176 us |       - |      - |       0 B |
|       BclListForEach | 22.326 us | 0.0120 us | 0.0100 us |       - |      - |       0 B |