@lahma 2019-01-09
Yes, good points. I'm in no hurry with this and I can PR separately if you are fine with the idea of making your original code look non-elegant and low-level 😉 

> My fear is that JIT will tomorrow (e.g. de-virtualization and tiered-compilation are on the way) and some of the optimizations may become redundant or even ineffective.

It's getting better and some optimizations have already come obsolete, but it usually requires that you are running on (latest) .NET Core instead the full framework that I'm now using...