@atifaziz 2019-01-09
> but it usually requires that you are running on (latest) .NET Core instead the full framework that I'm now using...

And so one more reason to consider CoreRT?

> if you are fine with the idea of making your original code look non-elegant and low-level 

Fine by me (of course) but I am hoping we can make much more macro performance gains that might yield more significant results before needing to resort to very fine-tuning. Just as an example, if you have a [`JsonWriter`](https://github.com/sebastienros/esprima-dotnet/blob/808dc9074091f8027a9dc10e9696b7357557f672/src/Esprima/Utils/JsonWriter.cs) implementation that emits _binary_ JSON like [UBJSON](http://ubjson.org/) then you can save the AST somewhere. For scripts that haven't changed (based on some content hash), you can re-hydrate the AST instead of parsing, avoiding JIT-ing the entire parser code! Also, all arrays behind `List<>` instances will then be just the needed size instead of have tail excess from auto-growth. Another macro optimization I mentioned once is to reuse AST nodes once we can guarantee [their immutability](https://github.com/sebastienros/esprima-dotnet/issues/66) (and location info is not needed).