@atifaziz 2019-01-09
@sebastienros I'm not suggesting that such infrastructure it belongs in Esprima but Esprima can help enable such things on the user's end by giving an immutable AST as well as having abstractions such as `JsonWriter` that permit writing to a more efficient binary version of JSON.