@sebastienros 2019-01-09
Why not use any generic serializer in this case? Protobuf, ... I don't think we need to change the JsonWriter.