@atifaziz 2019-01-09
> Why not use any generic serializer in this case?

For simplicity and to stay in the spirit of portability (thinking ESTree and staying naturally within the domain of JavaScript).

Also on the performance point, re-using AST nodes to further reduce GC pressure could be naturally added (as a macro optimization) to Esprima. It's not something to user needs to add after.